+++ src/qldpc/circuits/memory/alpha_syndrome_test.py
