+++ src/qldpc/circuits/memory/syndrome_measurement_test.py
